Move tests for multiselect-dropdown in one file
